+++ crates/node/src/main.rs
@@ -13,7 +13,7 @@ use cli::{NetworkType, RunCmd};
 use mc_analytics::Analytics;
 use mc_block_import::BlockImporter;
 use mc_db::DatabaseService;
-use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool};
+use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool, MempoolLimits};
 use mc_rpc::providers::{AddTransactionProvider, ForwardToProvider, MempoolAddTxProvider};
 use mc_telemetry::{SysInfo, TelemetryService};
 use mp_convert::ToFelt;
@@ -111,7 +111,11 @@ async fn main() -> anyhow::Result<()> {
     let l1_data_provider: Arc<dyn L1DataProvider> = Arc::new(l1_gas_setter.clone());
 
     // declare mempool here so that it can be used to process l1->l2 messages in the l1 service
-    let mempool = Arc::new(Mempool::new(Arc::clone(db_service.backend()), Arc::clone(&l1_data_provider)));
+    let mempool = Arc::new(Mempool::new(
+        Arc::clone(db_service.backend()),
+        Arc::clone(&l1_data_provider),
+        MempoolLimits::new(&chain_config),
+    ));
 
     let l1_service = L1SyncService::new(
         &run_cmd.l1_sync_params,